Move file out of utils.ts
